Add common films

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,23 +1,22 @@
 package ru.yandex.practicum.filmorate.service;
 
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import ru.yandex.practicum.filmorate.dal.DirectorRepository;
 import ru.yandex.practicum.filmorate.dal.FilmRepository;
+import ru.yandex.practicum.filmorate.exceptions.ConditionsNotMetException;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.enums.FeedEventOperation;
-import ru.yandex.practicum.filmorate.validator.DirectorValidator;
 import ru.yandex.practicum.filmorate.validator.FilmValidator;
 import ru.yandex.practicum.filmorate.validator.UserValidator;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class FilmService {
@@ -25,8 +24,9 @@ public class FilmService {
 	private final FilmValidator filmValidator;
 	private final UserValidator userValidator;
 	private final FeedService feedService;
-	private final DirectorValidator directorValidator;
-	private final DirectorService directorService;
+	private final DirectorRepository directorRepository;
+
+	public static final int CINEMA_BIRTH_YEAR = 1895;
 
 	public Film getFilmById(long filmId) {
 		return filmRepository.getFilmById(filmId)
@@ -53,6 +53,15 @@ public class FilmService {
 	}
 
 	public Collection<Film> mostPopular(Integer count, Integer year, Integer genreId) {
+		if (count != null && count < 0) {
+			throw new ConditionsNotMetException("Нужное количество не может быть отрицательным");
+		}
+		if (genreId != null && genreId < 0) {
+			throw new ConditionsNotMetException("Айди жанра не может быть отрицательным");
+		}
+		if (year != null && year < CINEMA_BIRTH_YEAR) {
+			throw new ConditionsNotMetException("Год не может быть раньше появления кинематографа");
+		}
 		return filmRepository.mostPopular(count, year, genreId);
 	}
 
@@ -61,7 +70,6 @@ public class FilmService {
 	}
 
 	public Film createFilm(Film film) {
-		log.info("Вызван метод createFilm с фильмом: {}", film);
 		List<Genre> genresList = checkFilmHasDuplicatedGenres(film);
 		film.setGenres(genresList);
 		filmValidator.validate(film);
@@ -76,22 +84,27 @@ public class FilmService {
 	}
 
 	public void deleteFilm(Long filmId) {
+		if (!filmValidator.filmExists(filmId)) {
+			throw new NotFoundException("Фильм id:" + filmId + " не найден");
+		}
 		filmRepository.deleteFilm(filmId);
 	}
 
 	public Collection<Film> getCommonFilms(Long userId, Long friendId) {
+		userValidator.userExists(userId);
+		userValidator.userExists(friendId);
 		return filmRepository.getCommonFilms(userId, friendId);
 	}
 
 	public List<Film> getFilmsByDirectorSortedByLikes(long directorId) {
-		Director director = directorService.getDirectorById(directorId);
-		directorValidator.validate(director);
+		if (!directorRepository.directorExists(directorId)) {
+			throw new NotFoundException("Такого режиссера у нас нет");
+		}
 		return filmRepository.getFilmsByDirectorSortedByLikes(directorId);
 	}
 
 	public List<Film> getFilmsByDirectorSortedByYears(long directorId) {
-		Director director = directorService.getDirectorById(directorId);
-		directorValidator.validate(director);
+		directorRepository.directorExists(directorId);
 		return filmRepository.getFilmsByDirectorSortedByYears(directorId);
 	}
 
@@ -112,6 +125,30 @@ public class FilmService {
                 .stream()
                 .distinct()
 				.toList();
+	}
+
+	public List<Film> searchFilms(String query, String by) {
+
+		List<String> search = Arrays.asList(by.split(","));
+		String title = "title";
+		String director = "director";
+
+		if (search.size() == 1 && search.contains(director)) {
+			return getFilmByDirector(query);
+		} else 	if (search.size() == 1 && search.contains(title)) {
+			return getFilmByTitle(query);
+		} else if (search.size() == 2 && search.contains(title) && search.contains(director)) {
+			return searchAll(query);
+		} else {
+			throw new ConditionsNotMetException("Неверные параметры поиска");
+		}
+	}
+
+	public List<Film> getFilmsByDirector(Long directorId, String sortBy) {
+		if (sortBy.equals("year")) {
+			return getFilmsByDirectorSortedByYears(directorId);
+		}
+		return getFilmsByDirectorSortedByLikes(directorId);
 	}
 }
